requirements.txt
asgiref==3.8.1
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
cloudinary==1.44.0
colorama==0.4.6
diff-match-patch==20241021
dj-database-url==2.3.0
Django==5.2
django-cloudinary-storage==0.3.0
django-import-export==4.3.7
django-jazzmin==3.0.1
django-storages==1.14.6
gunicorn==23.0.0
h11==0.16.0
idna==3.10
packaging==25.0
pillow==11.1.0
psycopg2-binary==2.9.10
python-decouple==3.8
python-dotenv==1.1.0
requests==2.32.3
six==1.17.0
sqlparse==0.5.3
tablib==3.8.0
typing_extensions==4.13.2
tzdata==2025.2
urllib3==2.4.0
uvicorn==0.34.2
whitenoise==6.9.0
